JSON Parser Added

+++ app/src/main/java/com/Example/iJam/myAdapter.java
@@ -46,7 +46,7 @@ public class myAdapter extends ArrayAdapter<myInterface> {
 
         txtTitle.setText(item.getTitle());
         txtLikes.setText(Float.toString(item.getLikes()));
-        rateBar.setRating(item.getRating());
+        rateBar.setRating((float)item.getRating());
 
         rowView.setOnClickListener(new View.OnClickListener(){
             @Override
